+++ testing/test_sample.py
@@ -60,7 +60,13 @@ class TestSamplePlot:
     def test_plot_sample(self, get_sample):
         plot_params = ins.signature(get_sample.plot).parameters["plot"].default
         plot_opts = get_args(plot_params)
-        assert all(isinstance(get_sample.plot(plot), mpl.axes.Axes) for plot in plot_opts) 
+        for plot in plot_opts:
+            if plot not in ["IR_to_DTG", "calibration", "fit"]:
+                assert isinstance(get_sample.plot(plot), mpl.axes.Axes) 
+            elif plot=="fit":
+                assert get_sample.plot(plot) == None
+            else:
+                assert get_sample.plot(plot) == True
 
 @pytest.mark.usefixtures("get_cali_worklist", "get_sample")
 #@pytest.mark.parametrize("get_sample, get_cali_worklist", [((sample_names[0], "Netzsch"), cali_sample_names)], indirect=True)